+++ scout/adapter/client.py
@@ -49,9 +49,7 @@ def get_connection(
             uri = "mongodb://{}:{}@{}:{}/{}".format(
                 quote_plus(username), quote_plus(password), host, port, authdb
             )
-            log_uri = "mongodb://{}:****@{}:{}/{}".format(
-                quote_plus(username), host, port, authdb
-            )
+            log_uri = "mongodb://{}:****@{}:{}/{}".format(quote_plus(username), host, port, authdb)
         else:
             log_uri = uri = "mongodb://%s:%s" % (host, port)