+++ scout/adapter/mongo/hgnc.py
@@ -75,9 +75,7 @@ class GeneHandler(object):
         # Add the transcripts:
         transcripts = []
         tx_objs = self.transcripts(build=build, hgnc_id=gene_obj["hgnc_id"])
-        nr_tx = sum(
-            1 for i in self.transcripts(build=build, hgnc_id=gene_obj["hgnc_id"])
-        )
+        nr_tx = sum(1 for i in self.transcripts(build=build, hgnc_id=gene_obj["hgnc_id"]))
         if nr_tx > 0:
             for tx in tx_objs:
                 transcripts.append(tx)
@@ -132,10 +130,7 @@ class GeneHandler(object):
                 return self.hgnc_collection.find(query)
 
             return self.hgnc_collection.find(
-                {
-                    "aliases": {"$regex": hgnc_symbol, "$options": "i"},
-                    "build": str(build),
-                }
+                {"aliases": {"$regex": hgnc_symbol, "$options": "i"}, "build": str(build),}
             )
 
         return self.hgnc_collection.find({"build": build, "aliases": hgnc_symbol})
@@ -275,13 +270,8 @@ class GeneHandler(object):
 
         res = self.hgnc_collection.find({"hgnc_symbol": symbol, "build": str(build)})
 
-        if (
-            self.hgnc_collection.find_one({"aliases": symbol, "build": str(build)})
-            is None
-        ):
-            LOG.debug(
-                "No gene with symbol {} was found. Attempting an alias.".format(symbol)
-            )
+        if self.hgnc_collection.find_one({"aliases": symbol, "build": str(build)}) is None:
+            LOG.debug("No gene with symbol {} was found. Attempting an alias.".format(symbol))
             res = self.hgnc_collection.find({"aliases": symbol, "build": str(build)})
 
         return res