+++ scout/commands/view/hgnc.py
@@ -10,11 +10,7 @@ LOG = logging.getLogger(__name__)
 @click.option("--hgnc-symbol", "-s", help="A valid hgnc symbol")
 @click.option("--hgnc-id", "-i", type=int, help="A valid hgnc id")
 @click.option(
-    "--build",
-    "-b",
-    type=click.Choice(["37", "38"]),
-    default="37",
-    help="Specify the genome build",
+    "--build", "-b", type=click.Choice(["37", "38"]), default="37", help="Specify the genome build",
 )
 @with_appcontext
 def hgnc(hgnc_symbol, hgnc_id, build):
@@ -41,9 +37,7 @@ def hgnc(hgnc_symbol, hgnc_id, build):
     click.echo("#hgnc_id\thgnc_symbol\taliases")
     for i, gene in enumerate(result, 1):
         click.echo(
-            "{0}\t{1}\t{2}".format(
-                gene["hgnc_id"], gene["hgnc_symbol"], ", ".join(gene["aliases"])
-            )
+            "{0}\t{1}\t{2}".format(gene["hgnc_id"], gene["hgnc_symbol"], ", ".join(gene["aliases"]))
         )
 
     if i == 0: