+++ scout/demo/__init__.py
@@ -32,19 +32,13 @@ clinical_snv_path = pkg_resources.resource_filename("scout", clinical_snv_file)
 clinical_sv_path = pkg_resources.resource_filename("scout", clinical_sv_file)
 clinical_str_path = pkg_resources.resource_filename("scout", clinical_str_file)
 
-customannotation_snv_path = pkg_resources.resource_filename(
-    "scout", customannotation_snv_file
-)
-vep_97_annotated_path = pkg_resources.resource_filename(
-    "scout", vep_97_annotated_snv_file
-)
+customannotation_snv_path = pkg_resources.resource_filename("scout", customannotation_snv_file)
+vep_97_annotated_path = pkg_resources.resource_filename("scout", vep_97_annotated_snv_file)
 
 research_snv_path = pkg_resources.resource_filename("scout", research_snv_file)
 research_sv_path = pkg_resources.resource_filename("scout", research_sv_file)
 
-cancer_sv_path = pkg_resources.resource_filename(
-    "scout", manta_annotated_sv_cancer_file
-)
+cancer_sv_path = pkg_resources.resource_filename("scout", manta_annotated_sv_cancer_file)
 
 empty_sv_clinical_path = pkg_resources.resource_filename("scout", empty_sv_file)