+++ scout/server/blueprints/cases/forms.py
@@ -33,9 +33,7 @@ class TagListField(Field):
 class GeneVariantFiltersForm(FlaskForm):
     """Base FiltersForm for SNVs"""
 
-    variant_type = SelectMultipleField(
-        choices=[("clinical", "clinical"), ("research", "research")]
-    )
+    variant_type = SelectMultipleField(choices=[("clinical", "clinical"), ("research", "research")])
     hgnc_symbols = TagListField("HGNC Symbols/Ids (case sensitive)")
     filter_variants = SubmitField(label="Filter variants")
     rank_score = IntegerField()