+++ tests/conftest.py
@@ -553,9 +553,7 @@ def adapter(request, pymongo_client):
 
 
 @pytest.fixture(scope="function")
-def clinvar_database(
-    request, adapter, clinvar_variant, user_obj, institute_obj, case_obj
-):
+def clinvar_database(request, adapter, clinvar_variant, user_obj, institute_obj, case_obj):
     "Returns an adapter to a database populated with one variant"
 
     user_id = user_obj["_id"]
@@ -708,11 +706,7 @@ def variant_database(request, populated_database):
     case_obj = adapter.case_collection.find_one()
 
     adapter.load_variants(
-        case_obj,
-        variant_type="clinical",
-        category="snv",
-        rank_threshold=-10,
-        build="37",
+        case_obj, variant_type="clinical", category="snv", rank_threshold=-10, build="37",
     )
 
     return adapter
@@ -727,11 +721,7 @@ def real_variant_database(request, real_populated_database):
     case_obj = adapter.case_collection.find_one()
     # Load variants
     adapter.load_variants(
-        case_obj,
-        variant_type="clinical",
-        category="snv",
-        rank_threshold=-10,
-        build="37",
+        case_obj, variant_type="clinical", category="snv", rank_threshold=-10, build="37",
     )
 
     return adapter
@@ -1405,13 +1395,8 @@ def mme_submission():
 @pytest.fixture(scope="function")
 def mme_patient():
     json_patient = {
-        "contact": {
-            "href": "mailto:contact_email@email.com",
-            "name": "A contact at an institute",
-        },
-        "features": [
-            {"id": "HP:0001644", "label": "Dilated cardiomyopathy", "observed": "yes"}
-        ],
+        "contact": {"href": "mailto:contact_email@email.com", "name": "A contact at an institute",},
+        "features": [{"id": "HP:0001644", "label": "Dilated cardiomyopathy", "observed": "yes"}],
         "genomicFeatures": [
             {
                 "gene": {"id": "LIMS2"},