+++ CHANGELOG.md
@@ -6,7 +6,9 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 
 ## [x.x.x]
 ### Added
+- COSMIC badge shown in cancer variants
 - Mark MANE transcript in list of transcripts in "Transcript overview" on variant page
+
 
 ### Fixed
 - Bug in clinVar form when variant has no gene